Rename charts folder to dashboard, move DreamDashboard into renamed folder
